EY-2724 - Delt tilbakekreving i komponenter frontend og tilpasset navngivning i domene
- Ny komponent TilbakekrevingVurderingOverordnet og TilbakekrevingVurderingPerioder
- Renamet ytelsebeloeper -> ytelse
